// src/api/models/dashboard-summary-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { DashboardActivityDto } from '../models/dashboard-activity-dto';
import { DashboardTrendBarDto } from '../models/dashboard-trend-bar-dto';
export interface DashboardSummaryDto {
  activeLoanPct?: number;
  collectedAmount?: number;
  collectionProgress?: number;
  npaLoanPct?: number;
  overdueLoanPct?: number;
  recentActivities?: Array<DashboardActivityDto>;
  shortfall?: number;
  todaysTarget?: number;
  totalActiveLoans?: number;
  totalCustomers?: number;
  totalPortfolioValue?: number;
  trendBars?: Array<DashboardTrendBarDto>;
}